Extract HTTP REST resource creation method
